flt2dec/random.rs renamed to flt2dec/equiv.rs
