+++ library/coretests/tests/num/flt2dec/equiv.rs
